cmd: adds top-level account command
