+++ core/api_account.go
@@ -1,0 +1,11 @@
+package core
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func (a *api) accountAddress(g *gin.Context) {
+	g.String(http.StatusOK, a.node.account.Address())
+}
